Fix some e2e tests
