+++ packages/components/src/guide/index.js
@@ -64,6 +64,20 @@ export default function Guide( {
 		}
 	};
 
+	useEffect( () => {
+		// Keeping the focus within the guide when the page changes
+		// prevents the modal from closing and avoids focus loss.
+		if (
+			guideContainer.current.contains(
+				guideContainer.current.ownerDocument.activeElement
+			)
+		) {
+			return;
+		}
+
+		focus.tabbable.find( guideContainer.current )?.[ 0 ]?.focus();
+	}, [ currentPage ] );
+
 	if ( pages.length === 0 ) {
 		return null;
 	}
@@ -82,7 +96,7 @@ export default function Guide( {
 			} }
 			ref={ guideContainer }
 		>
-			<div className="components-guide__container">
+			<div className="components-guide__container" ref={ guideContainer }>
 				<div className="components-guide__page">
 					{ pages[ currentPage ].image }
